Speech to Text with Whisper
The following branch is meant to be dedicated to following Aadi's advice on how to approach the project. The OpenAI Whisper model is tested for converting the audio file to text. Currently, we should aim to setup an joint account of sorts which has premium access to Open AI's resources.